adapt template to not add empty address line
if network facts are used to specify ignored ip addresses,
and puppet is used to configure the network interfaces the
ip address fact only available after the first successful
puppet run (or even a reboot, if you do not reconfigure
network interfaces with puppet).
Currently the situation leads to an incorrect conntrackd
configuration which fails starting conntrackd.
This patch ignores the empty address fields so conntrackd
is not failing after first run.
(reconfiguration in second run is still guaranted)